requirements.txt
numpy==1.26.4
matplotlib==3.10.0
torch==2.2.2+cpu
opencv-python==4.10.0
Pillow==10.2.0
facenet-pytorch